+++ client/src/components/outer-components/result-types/WebResults.jsx
@@ -22,6 +22,7 @@ const WebResults = ({ webResults, onDdgClick }) => {
                   Learn more
                 </a>
               )}
+              
             </li>
           ))}
       </ul>
